"Refactor SQLite pool function and improve function naming
The SQLite pool function has been moved to a dedicated SQLite module for better code organization. Multiple function names have been adjusted for more accurate descriptions. This contributes to clearer code structure and improved readability."
