+++ src/cmd/anilist_module/user.rs
@@ -1,5 +1,5 @@
 use crate::function::error_management::common::custom_error;
-use crate::function::sql::sql::get_sqlite_pool;
+use crate::function::sql::sqlite::pool::get_sqlite_pool;
 use crate::structure::anilist::user::struct_user::UserWrapper;
 use crate::structure::embed::anilist::struct_lang_user::UserLocalisedText;
 use crate::structure::register::anilist::struct_user_register::RegisterLocalisedUser;
